requirements.txt
fastapi
uvicorn[standard]
pydantic
trafilatura
google-genai
python-dotenv
beautifulsoup4
requests
